+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-07-25] [rework scoped shape, support context in drill](https://github.com/RubricLab/chains/commit/fdd44bba0accb530d653c4fa72c1b9344b550633)
 - [2025-07-18] [cleanup](https://github.com/RubricLab/chains/commit/044d6c94f1b20a6317679f45a247c690a3c1e4db)
 - [2025-07-18] [cleanup, use @/shapes](https://github.com/RubricLab/chains/commit/4cdfa583df69796db9796d555e51e2bf87a67a57)
 - [2025-07-17] [scoped, branded, custom major build](https://github.com/RubricLab/chains/commit/d2c1f59ed25ccd7e42c1bf809d6ac7beb3f9543f)